+++ CHANGELOG.md
@@ -13,8 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Add dependency "[cake-case](https://github.com/archco/cake-case#readme)". #103
-- Add style for `<kbd>` tag. #109
+- Added dependency "[cake-case](https://github.com/archco/cake-case#readme)". #103
+- Added style for `<kbd>` tag. #109
+- Added scss-class `.btn-text`. #113
 
 ### Changed