+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][]
 
+## [0.6.0][] - 2018-04-22
+
 ### Added
 
 - Added dependency "[cake-case](https://github.com/archco/cake-case#readme)". #103
@@ -207,7 +209,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove unnecessary part "aside.scss". #60
 - Remove deprecated scss mixin, `flex-container()`. instead use `flex-container-align`. #52
 
-[Unreleased]: https://github.com/archco/moss-ui/compare/v0.5.3...HEAD
+[Unreleased]: https://github.com/archco/moss-ui/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/archco/moss-ui/compare/v0.5.3...v0.6.0
 [0.5.3]: https://github.com/archco/moss-ui/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/archco/moss-ui/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/archco/moss-ui/compare/v0.5.0...v0.5.1